Create NewYorkCityBlog
This second project is part of the learning journey with Codecademy with learning the fundamentals of HTML
 Project Description: 
After researching New York City, you decide to create a blog for your viewers who want to know more about the city. It’s time to create a blog to show off how amazing the Big Apple is. You got this!"